requirements.txt
faiss-cpu
langchain
langchain-community
langchain-huggingface
openai
python-dotenv
streamlit
torch